colocate plot context stuff
