+++ src/components/BarSeries.tsx
@@ -1,6 +1,6 @@
 import { CSSProperties, useEffect, useMemo, useState } from 'react';
 
-import { usePlotContext } from '../hooks';
+import { usePlotContext } from '../plotContext';
 import type { SeriesPoint } from '../types';
 import { getNextId, validateAxis } from '../utils';
